refactor floating panes out of tab
