+++ zellij-server/src/panes/mod.rs
@@ -5,6 +5,7 @@ mod plugin_pane;
 mod selection;
 pub mod terminal_character;
 mod terminal_pane;
+mod floating_panes;
 
 pub use alacritty_functions::*;
 pub use grid::*;
@@ -12,3 +13,4 @@ pub use link_handler::*;
 pub(crate) use plugin_pane::*;
 pub(crate) use terminal_character::*;
 pub use terminal_pane::*;
+pub use floating_panes::*;
